+++ pages/options.js
@@ -1,3 +1,23 @@
+/*
+    options.js is part of anti-social, a Firefox extension to help users
+    stop themselves from wasting too much time on social media.
+
+    Copyright (C) 2018  Wayne Myers (wgmyers@gmail.com)
+
+    anti-social is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
 "use strict";
 
 // Default settings object